splash screen added blanc
