+++ PEP_Android/app/src/main/AndroidManifest.xml
@@ -8,18 +8,24 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".Activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
+        </activity>
+        <activity android:name=".Activity.ScannerActivity" />
+        <activity android:name=".Activity.PassportActivity" />
+        <activity android:name=".Activity.SettingsActivity" />
+        <activity android:name=".Activity.StijlActivity" />
+        <activity android:name=".Activity.KleurenPaletActivity" />
+        <activity android:name=".Activity.Splash">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ScannerActivity" />
-        <activity android:name=".PassportActivity" />
-        <activity android:name=".SettingsActivity" />
-        <activity android:name=".StijlActivity" />
-        <activity android:name=".KleurenPaletActivity" />
     </application>
 
 </manifest>
