+++ PEP_Android/app/src/main/java/com/avans/in11sob/pep_android/Activity/KleurenPaletActivity.java
@@ -1,0 +1,15 @@
+package com.avans.in11sob.pep_android.Activity;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.avans.in11sob.pep_android.R;
+
+public class KleurenPaletActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_kleuren_palet);
+    }
+}
