+++ PEP_Android/app/src/main/java/com/avans/in11sob/pep_android/Activity/ScannerActivity.java
@@ -1,7 +1,9 @@
-package com.avans.in11sob.pep_android;
+package com.avans.in11sob.pep_android.Activity;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.avans.in11sob.pep_android.R;
 
 public class ScannerActivity extends AppCompatActivity {
