+++ PEP_Android/app/src/main/java/com/avans/in11sob/pep_android/Activity/Splash.java
@@ -1,0 +1,15 @@
+package com.avans.in11sob.pep_android.Activity;
+
+import android.app.Activity;
+import android.os.Bundle;
+
+import com.avans.in11sob.pep_android.R;
+
+public class Splash extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.splash);
+    }
+}
