--- PEP_Android/app/src/main/java/com/avans/in11sob/pep_android/SettingsActivity.java
@@ -1,13 +1,0 @@
-package com.avans.in11sob.pep_android;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class SettingsActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_settings);
-    }
-}
