+++ PEP_Android/app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.avans.in11sob.pep_android.MainActivity"
+    tools:context=".Activity.MainActivity"
     android:weightSum="1">
 
     <Button
