+++ PEP_Android/app/src/main/res/layout/activity_passport.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.avans.in11sob.pep_android.PassportActivity">
+    tools:context=".Activity.PassportActivity">
 
     <Button
         android:id="@+id/stijl"
