+++ PEP_Android/app/src/main/res/layout/activity_settings.xml
@@ -7,6 +7,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.avans.in11sob.pep_android.SettingsActivity">
+    tools:context=".Activity.SettingsActivity">
 
 </RelativeLayout>
